feat: search by interaction id, perf opt.
