+++ src/gateway/router/gatewayRouter.ts
@@ -1,12 +1,12 @@
 import Router from "@koa/router";
 import {contractsRoute} from "./routes/contractsRoute";
 import {interactionsRoute} from "./routes/interactionsRoute";
-import {searchContractsRoute} from "./routes/searchContractsRoute";
+import {searchRoute} from "./routes/searchRoute";
 
 const gatewayRouter = new Router({prefix: '/gateway'});
 
 gatewayRouter.get("/contracts", contractsRoute);
-gatewayRouter.get("/contracts/search/:phrase", searchContractsRoute);
+gatewayRouter.get("/search/:phrase", searchRoute);
 gatewayRouter.get("/interactions", interactionsRoute);
 
 export default gatewayRouter;
